Fix #1261: Modified error_get_last output elements to be comaptible with zend
This fixed issue #1261

I changed BaseExecutionContext::recordLastError() to be virtual function, and override it in VMExecutionContext class.
The reason why I override it is that VMExecutionContext is accessible for file path and line number, whici don't accessible by BaseExecutionContext.

Although it seems that recordLastError() would be a bit slower due to change to be virtual function, I adopted this method.
Because recordLastError() is called when some error is occurred, I guess this method don't affect a performance of hhvm very much..