Detect cycles only when necessary
Move the responsibility for cycle detection from blockOn() to its
caller. Do not detect cycles when the parent wait handle is being
creates, as it's impossible for new cycle to be introduced that way.

Also, it was previously not possible to catch a cycle exception by
adding try/catch around the await. Fix it.

Reviewed By: @jdelong

Differential Revision: D1062770