Some improvements to the iter_next "cold" path
Most of the time when we go to the cold path, it's because
we're iterating an APCLocalArray.  Add a specialized version for that.
We were also always going to cold by first going to a function that
checked for various collection types (with very big code)---change
this to a quick check at the front of iter_next_cold that then tail
calls to a function that code collection dispatch on the already
loaded type (compiles as a jump table).  We also bailed for releasing
the array which is easy enough to handle.

Reviewed By: @bertmaher

Differential Revision: D1067317