Maintain in-generator state, remove dependency on func->isGenerator()
Add a flag to ActRec indicating whether the ActRec is stored inside a
Continuation object. Maintain in-generator state in the HHBC translator
and use it instead of curFunc->isGenerator() checks.

Makes it possible to unify inner and outer generator functions.

All remaining func->isGenerator() checks are directly related to the
inner/outer split and will go away in a diff that merges them.

Reviewed By: @jdelong

Differential Revision: D1095272