Don't emit type predictions in TransOptimize mode
I noticed that the region translator was emitting type checks for
values that were both predicted (via the interpreter type profiler)
and inferred (via the front-end).  This is a result of the type
profiler annotating the output of the instructions while the front-end
annotates the instruction inputs.  Therefore, we ended up generating a
CheckStk instruction and then an AssertStk instruction.  The AssertStk
instruction is then simplified away (since it's asserting a type that
is already known due to the previous CheckStk), and the CheckStk
remains. The tracelet translator avoids this problem by looking ahead
in the chain of NormalizedInstructions (via getOutputUsage) to
determine that the prediction is unecessary.

I PGO mode, we don't really need the values coming from the
interpreter type profiler.  TransProfile translations end whenever
there's a side-exit, and type predictions incur side-exits.  And when
we stitch multiple TransProfile translations together to for a
hottrace region (in TransOptimize mode), the guard for the top of the
stack essentially does the role of type prediction.  And, if the value
is also inferred, then the guard is omitted and the AssertStk is
emitted and not eliminated (since there's not previous CheckStk).

Note that this diff only solves the problem for regions formed via PGO
(hottrace only).  Fixing the problem for other region selectors is
going to be more involved and I tasked to do it separately.

Reviewed By: @swtaarrs

Differential Revision: D1112059