fix constructor param for FilterIterator
When grabbing the constructor arguments through reflection from the
FilterIterator class, HHVM accurately reports the parameter with name
'it' in accordance to both HHVM's FilterIterator.php and PHP's
filteriterator.inc(1).

PHP however, returns 'iterator'. I suspect this is a core PHP bug but I
am running this through you guys to see if this should be reported
upstream or if I am just imagining things.

For your information, the classes that do derive from FilterIterator
already take an 'iterator' param. In any case, attached is a patch
and testcase to make HHVM compatible with PHP, despite it doing crazy stuff.

1) https://github.com/php/php-src/blob/master/ext/spl/internal/filteriterator.inc#L35